Add files via upload
Initial commit for basic Pie Chart Functionality

This commit shows the basic idea of how my pie chart will work. It will take in the monthly income and desired budget category percentages and then display a pie chart and the actual monetary values that the percentages imply. As of this commit, it does not actually take in input, it is just set to a random set of values. Later I will create the HTML file that can place and handle the input text boxes required.